Replace `init` with `yo`
