+++ packages/cli/docs/cli.html
@@ -256,7 +256,6 @@ a code {
 <li><a href="#envset">env:set</a></li>
 <li><a href="#envunset">env:unset</a></li>
 <li><a href="#history">history</a></li>
-<li><a href="#init">init</a></li>
 <li><a href="#integrations">integrations</a></li>
 <li><a href="#link">link</a></li>
 <li><a href="#login">login</a></li>
@@ -278,7 +277,6 @@ a code {
 <li><a href="#usersremove">users:remove</a></li>
 <li><a href="#validate">validate</a></li>
 <li><a href="#versions">versions</a></li>
-<li><a href="#yo">yo</a></li>
 </ul>
 </li>
 </ul>
@@ -610,34 +608,6 @@ a code {
 </blockquote><p><strong>Usage</strong>: <code>zapier history</code></p><p>History includes all the changes made over the lifetime of your integration. This includes everything from creation, updates, migrations, admins, and invitee changes, as well as who made the change and when.</p><p><strong>Flags</strong></p><ul>
 <li><code>-f, --format</code> | Change the way structured data is presented. If &quot;json&quot; or &quot;raw&quot;, you can pipe the output of the command into other tools, such as jq. One of <code>[plain | json | raw | row | table]</code>. Defaults to <code>table</code>.</li>
 <li><code>-d, --debug</code> | Show extra debugging output.</li>
-</ul>
-    </div>
-    <div class="col-md-7 col-sm-12 col-height is-empty docs-code">
-      
-    </div>
-  </div>
-</div><div class="row">
-  <div class="row-height">
-    <div class="col-md-5 col-sm-12 col-height  docs-primary">
-      <h2 id="init">init</h2>
-    </div>
-    <div class="col-md-7 col-sm-12 col-height is-empty docs-code">
-      
-    </div>
-  </div>
-</div><div class="row">
-  <div class="row-height">
-    <div class="col-md-5 col-sm-12 col-height  docs-primary">
-      <blockquote>
-<p>Initialize a new Zapier integration. Optionally uses a template.</p>
-</blockquote><p><strong>Usage</strong>: <code>zapier init PATH</code></p><p>After running this, you&apos;ll have a new example integration in your directory. If you re-run this command on an existing directory it will leave existing files alone and not clobber them.</p><p>This doesn&apos;t register or deploy the integration with Zapier - try the <code>zapier register</code> and <code>zapier push</code> commands for that!</p><p><strong>Arguments</strong></p><ul>
-<li>(required) <code>path</code> | Where to create the new integration. If the directory doesn&apos;t exist, it will be created. If the directory isn&apos;t empty, we&apos;ll ask for confirmation</li>
-</ul><p><strong>Flags</strong></p><ul>
-<li><code>-t, --template</code> | The template to start your integration with. One of <code>[minimal | trigger | search | create | basic-auth | custom-auth | digest-auth | oauth2 | oauth1-trello | oauth1-tumblr | oauth1-twitter | session-auth | dynamic-dropdown | files | middleware | resource | rest-hooks | search-or-create | babel | typescript | github | onedrive]</code>. Defaults to <code>minimal</code>.</li>
-<li><code>-d, --debug</code> | Show extra debugging output.</li>
-</ul><p><strong>Examples</strong></p><ul>
-<li><code>zapier init ./some/path</code></li>
-<li><code>zapier init . --template typescript</code></li>
 </ul>
     </div>
     <div class="col-md-7 col-sm-12 col-height is-empty docs-code">
@@ -1213,34 +1183,6 @@ a code {
 </blockquote><p><strong>Usage</strong>: <code>zapier versions</code></p><p><strong>Flags</strong></p><ul>
 <li><code>-f, --format</code> | Change the way structured data is presented. If &quot;json&quot; or &quot;raw&quot;, you can pipe the output of the command into other tools, such as jq. One of <code>[plain | json | raw | row | table]</code>. Defaults to <code>table</code>.</li>
 <li><code>-d, --debug</code> | Show extra debugging output.</li>
-</ul>
-    </div>
-    <div class="col-md-7 col-sm-12 col-height is-empty docs-code">
-      
-    </div>
-  </div>
-</div><div class="row">
-  <div class="row-height">
-    <div class="col-md-5 col-sm-12 col-height  docs-primary">
-      <h2 id="yo">yo</h2>
-    </div>
-    <div class="col-md-7 col-sm-12 col-height is-empty docs-code">
-      
-    </div>
-  </div>
-</div><div class="row">
-  <div class="row-height">
-    <div class="col-md-5 col-sm-12 col-height  docs-primary">
-      <blockquote>
-<p>Initialize a new Zapier integration. Optionally uses a template.</p>
-</blockquote><p><strong>Usage</strong>: <code>zapier yo PATH</code></p><p>After running this, you&apos;ll have a new integration in the specified directory. If you re-run this command on an existing directory, it will prompt before overwriting any existing files.</p><p>This doesn&apos;t register or deploy the integration with Zapier - try the <code>zapier register</code> and <code>zapier push</code> commands for that!</p><p><strong>Arguments</strong></p><ul>
-<li>(required) <code>path</code> | Where to create the new integration. If the directory doesn&apos;t exist, it will be created. If the directory isn&apos;t empty, we&apos;ll ask for confirmation</li>
-</ul><p><strong>Flags</strong></p><ul>
-<li><code>-t, --template</code> | The template to start your integration with. One of <code>[basic-auth | custom-auth | digest-auth | dynamic-dropdown | files | minimal | oauth1-trello | oauth2 | search-or-create | session-auth | typescript]</code>.</li>
-<li><code>-d, --debug</code> | Show extra debugging output.</li>
-</ul><p><strong>Examples</strong></p><ul>
-<li><code>zapier yo myapp</code></li>
-<li><code>zapier yo ./path/myapp --template oauth2</code></li>
 </ul>
     </div>
     <div class="col-md-7 col-sm-12 col-height is-empty docs-code">
